Added SKIP_SYSCLK_INIT for the loader
In general the Loader will switch off the SDRAM clock at start. This is a problem in case of debugging a flash session where the SDRAM clock was initialize before by the startup script. This can be prevented by using the SKIP_SYSCLK_INIT switch. However, it must be ensure that the clock is initialized before the Loader using the startup script.