Added small delay in case of a octo flash
I use a RT1062 CPU with octo flash.

When debugging in Flash the download only works every second time for me. The
error message is:

Erase failed
Memory erase operation failed: no driver installed for memory range

But when the loader is compiled with debug output, it works here every time.
It looks like a small delay is needed to make the download work every time.